requirements.txt
streamlit
pandas
numpy
openpyxl
Pillow
requests
pycryptodome
torch
torchvision
